// src/generated/material-api/model/speakingSectionSummaryDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { SpeakingSectionSummaryDtoStatus } from './speakingSectionSummaryDtoStatus';

export interface SpeakingSectionSummaryDto {
  materialId: number;
  sectionId: number;
  sectionTitle: string;
  part1Id: number;
  part1Title: string;
  part2Id: number;
  part2Title: string;
  status: SpeakingSectionSummaryDtoStatus;
  createdAt: string;
  updatedAt: string;
}