// src/generated/material-api/model/speakingSectionEditDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { SpeakingSectionEditDtoStatus } from './speakingSectionEditDtoStatus';
import type { QuestionEditDto } from './questionEditDto';

export interface SpeakingSectionEditDto {
  materialId: number;
  sectionId: number;
  status: SpeakingSectionEditDtoStatus;
  materialTitle: string;
  materialDescription: string;
  partTitle: string;
  partImageStorageKey: string;
  questions: QuestionEditDto[];
  part2Title: string;
  part2Questions: QuestionEditDto[];
}