// src/generated/material-api/model/materialNodeWithAssetsDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MaterialNodeWithAssetsDtoKind } from './materialNodeWithAssetsDtoKind';
import type { MaterialAssetDto } from './materialAssetDto';
import type { MaterialNodeWithAssetsDtoConfig } from './materialNodeWithAssetsDtoConfig';

export interface MaterialNodeWithAssetsDto {
  id: number;
  parentNodeId: number;
  /** Node type within the material tree */
  kind: MaterialNodeWithAssetsDtoKind;
  title: string;
  displayOrder: number;
  transcriptText: string;
  assets: MaterialAssetDto[];
  /** Parsed config object for this node. */
  config: MaterialNodeWithAssetsDtoConfig;
}